+++ src/modules/contact/contact.service.ts
@@ -34,6 +34,10 @@ export class ContactService {
     return this.getEngine(sessionId).checkNumberExists(number);
   }
 
+  getNumberId(sessionId: string, number: string) {
+    return this.getEngine(sessionId).getNumberId(number);
+  }
+
   getProfilePicture(sessionId: string, contactId: string) {
     return this.getEngine(sessionId).getProfilePicture(contactId);
   }